Server & client side validation update for rent apartment form

+++ database/migrations/2023_11_27_000613_create_listings_table.php
@@ -19,9 +19,9 @@ return new class extends Migration
             $table->string('property_type', 30);
             $table->string('location');
             $table->string('street');
-            $table->decimal('price', 10, 2);
-            $table->decimal('property_area', 8, 2);
-            $table->decimal('land_area', 8, 2)->nullable();
+            $table->integer('price');
+            $table->integer('property_area');
+            $table->integer('land_area')->nullable();
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users');
